Removed DataList, DataModel and Query. Moved request to api folder.
