+++ src/views/Samples.vue
@@ -24,7 +24,7 @@ import Alert from '@/components/Alert'
 import ApiConfig from '@/api/Config'
 import ConfigItem from '@/api/ConfigItem'
 import ComponentFactory from '@/mixins/ComponentFactory'
-import Request from '@/mixins/Request'
+import Request from '@/api/Request'
 import Response from '@/api/Response'
 import Modal from '@/components/Modal'
